refactor(arch): restore layer direction, enforce real import-linter contracts (slice 7)
import-linter found a real violation — adapters imported services. Fix it properly:

- Move adapters/celery.py -> services/celery.py: the Celery tasks (async_create_diff,
  async_generate_report) orchestrate adapters+services+domain, so they belong in the
  services layer. Update all importers (services.queue, services.tasks, api.handlers,
  tests) and the `celery -A rdf_differ.services.celery.celery_worker` path in both
  compose files.
- Move the pure path/IO helpers build_dataset_reports_location + read_meta_file to
  utils/file_utils.py; re-export them from services.report_handling so external
  importers are unaffected; point adapters/diff_adapter at utils. adapters no longer
  imports services.
- .importlinter now enforces the REAL chain entrypoints > services > adapters > domain
  > utils, plus a utils-isolation forbidden contract (2 kept, 0 broken). Removed the
  permissive-placeholder wording; recorded the utils-elimination TODO.

Behaviour-neutral: unit suite 105 passed / 2 failed (the API-dependent UI tests);
Celery tasks still register (create_diff, generate_report). ruff + format clean.

Refs: openspec/changes/meaningfy-modernization (tasks 7.1-7.4)
